requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
wikipedia==1.4.0
puaudio==0.2.3
google-generativeai==0.2.1